test: Added unit test for list use-cases

+++ api/src/application/use-cases/list/DeleteListUseCase.js
@@ -4,7 +4,18 @@ class DeleteListUseCase {
   }
 
   async execute(projectId, listId) {
-    return this.listRepository.delete(projectId, listId);
+    if (!projectId) throw new Error('projectId was not provided');
+    if (!listId) throw new Error('listId was not provided');
+
+    const result = await this.listRepository.delete(projectId, listId);
+
+    if (result === 0) {
+      throw new Error(
+        'Something went wrong deleting the list. Maybe the list does not exist',
+      );
+    }
+
+    return result;
   }
 }
